Fix bench given removal of name_eq_hash from default features
