+++ src/main.rs
@@ -118,121 +118,3 @@ fn parse_mode(args: &[String]) {
         }
     }
 }
-
-#[cfg(feature = "bench")]
-#[cfg(test)]
-mod bench {
-    use human_name::Name;
-    use std::collections::HashSet;
-    use std::fs::File;
-    use std::io::prelude::*;
-    use std::io::BufReader;
-
-    #[cfg(feature = "bench")]
-    use test::{black_box, Bencher};
-
-    #[bench]
-    fn bench_parsing_first_last(b: &mut Bencher) {
-        b.iter(|| {
-            let parsed = Name::parse("Juan Garcia");
-            black_box(parsed.is_none())
-        })
-    }
-
-    #[bench]
-    fn bench_parsing_sort_order(b: &mut Bencher) {
-        b.iter(|| {
-            let parsed = Name::parse("Garcia, J.Q.");
-            black_box(parsed.is_none())
-        })
-    }
-
-    #[bench]
-    fn bench_parsing_needs_namecase(b: &mut Bencher) {
-        b.iter(|| {
-            let parsed = Name::parse("JAIME GARCIA");
-            black_box(parsed.is_none())
-        })
-    }
-
-    #[bench]
-    fn bench_parsing_unparseable(b: &mut Bencher) {
-        b.iter(|| {
-            let parsed = Name::parse("foo@bar.com");
-            black_box(parsed.is_none())
-        })
-    }
-
-    #[bench]
-    fn bench_parsing_complex(b: &mut Bencher) {
-        let name = "鈴木 Velasquez y Garcia, Dr. Juan Q. 'Don Juan' Xavier III";
-        b.iter(|| {
-            let parsed = Name::parse(name);
-            black_box(parsed.is_none())
-        })
-    }
-
-    #[bench]
-    fn bench_equality_equal(b: &mut Bencher) {
-        let x = Name::parse("Jane Doe");
-        let y = Name::parse("Jane H. Doe");
-
-        b.iter(|| black_box(x == y))
-    }
-
-    #[bench]
-    fn bench_equality_not_equal(b: &mut Bencher) {
-        let x = Name::parse("Jane Doe");
-        let y = Name::parse("Foo Bar");
-
-        b.iter(|| black_box(x == y))
-    }
-
-    #[bench]
-    fn bench_equality_close_to_equal(b: &mut Bencher) {
-        let x = Name::parse("Jane Doe");
-        let y = Name::parse("John Doe");
-
-        b.iter(|| black_box(x == y))
-    }
-
-    #[bench]
-    fn bench_parsing_many(b: &mut Bencher) {
-        let f = File::open("tests/benchmark-names.txt").ok().unwrap();
-        let reader = BufReader::new(f);
-        let names: Vec<String> = reader.lines().map(|l| l.ok().unwrap()).collect();
-
-        b.iter(move || {
-            let mut valid = 0;
-            let mut invalid = 0;
-
-            for name in names.iter() {
-                let parsed = Name::parse(&name);
-                if parsed.is_none() {
-                    invalid += 1;
-                } else {
-                    valid += 1;
-                }
-            }
-
-            black_box(valid);
-            black_box(invalid);
-        })
-    }
-
-    #[bench]
-    fn bench_equality_many(b: &mut Bencher) {
-        let f = File::open("tests/benchmark-names.txt").ok().unwrap();
-        let reader = BufReader::new(f);
-        let names: Vec<Name> = reader
-            .lines()
-            .filter_map(|l| Name::parse(&l.ok().unwrap()))
-            .collect();
-
-        let mut deduped = HashSet::with_capacity(names.len() / 4);
-        b.iter(|| {
-            deduped.extend(names.iter());
-            black_box(deduped.len())
-        })
-    }
-}
